// src/app/api/models/update-profile-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UpdateProfileDto {
  address?: string | null;
  address2?: string | null;
  city?: string | null;
  country?: string | null;
  dateOfBirth?: string | null;
  gender?: 'male' | 'female' | 'other' | 'prefer_not_to_say' | 'null';
  lastName?: string | null;
  name?: string;
  phone?: string | null;
  state?: string | null;
  username?: string | null;
  zipCode?: string | null;
}